Add `global-backup/password` to list of secrets to generate
Currently the Commodore component backup-k8up requires the secret
`global-backup/password` to be present for the cluster in Vault, if the
setting `global_backup_config.enabled` is `true`.

Since we use other parts of `global_backup_config` to provide sensible
defaults for VSHN-managed OpenShift4 on Cloudscale we have to set the
global-backup/password secret.