Use `${CLUSTER_DOMAIN}` variable instead of hardcoded cluster domain
Previously we used `${CLUSTER_ID}.${BASE_DOMAIN}` in some places in the
how-to. With adding support for customizing the cluster domain, these
need to be dynamically adjusted to use the correct cluster domain.

This commit replaces hardcoded cluster domain references with variable
`${CLUSTER_DOMAIN}` which contains the cluster domain extracted from the
manifests generated by the OpenShift installer.