Remove steps to manually deploy image registry S3 credentials
We no longer need to deploy the image registry S3 credentials (and the
newly introduced partial config) before synthesizing the cluster since
the registry operator won't try to use S3 storage on its own for UPI on
platform External as far as I can tell.

Instead we add a new section that verifies the state of the internal
registry after the cluster has been synthesized.

With this change, the install instructions expect that the cluster's
Project Syn config sets `storage.managementState=Unmanaged` in the
`config.imageregistry` custom resource.